+++ lib/br_invoices_pdf/cfe.rb
@@ -1,0 +1,8 @@
+require 'br_invoices_pdf/cfe/parser'
+require 'br_invoices_pdf/cfe/renderer'
+
+module BrInvoicesPdf
+  module Cfe
+    BrInvoicesPdf.register(:cfe, Cfe::Renderer, Cfe::Parser)
+  end
+end